Removing unused core data
