--- F1App/Entities/BoxBoxCoreData.xcdatamodeld/BoxBoxCoreData.xcdatamodel/contents
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="22225" systemVersion="23B2082" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
-    <entity name="UserEntity" representedClassName=".BoxBoxCoreData" syncable="YES" codeGenerationType="class">
-        <attribute name="favoriteDriver" optional="YES" attributeType="String"/>
-        <attribute name="favoriteTeam" optional="YES" attributeType="String"/>
-        <attribute name="imgData" optional="YES" attributeType="Binary" allowsExternalBinaryDataStorage="YES"/>
-        <attribute name="username" optional="YES" attributeType="String"/>
-    </entity>
-</model>
